new
Selection and implementation of a web-scale discovery tool by the Rider
University Libraries (RUL) in
the 2011–2012 academic year revealed that the endeavor was a complex
one. Research into the state
of adoption of web-scale discovery tools in North America and the
evolution of product effectiveness
provided a good starting point. In the following study, we evaluated
fourteen major discovery tools
(three open source and ten proprietary), benchmarking sixteen criteria
recognized as the advanced
features of a “next generation catalog.” Some of the features have been
used in previous research on
discovery tools. The purpose of the study was to evaluate and compare
all the major discovery tools,
and the findings serve to update librarians on the latest developments
and user interfaces and to
assist them in their adoption of a discovery tool.